Protect against interrupted download
